chore: renamed directory appInterface/settingsElements to appInterface/dialogElements
