+++ app/src/main/kotlin/view/appInterface/dialog/Settings.kt
@@ -25,14 +25,14 @@ import androidx.compose.ui.window.Dialog
 import org.coremapx.app.config
 import org.coremapx.app.localization.LocalizationManager
 import org.coremapx.app.theme.AppTheme
-import view.appInterface.settingsElements.blocks.ColorsBlock
-import view.appInterface.settingsElements.blocks.CommandFieldBlock
-import view.appInterface.settingsElements.blocks.GeneralBlock
-import view.appInterface.settingsElements.blocks.MainScreenBlock
-import view.appInterface.settingsElements.blocks.PerformanceBlock
-import view.appInterface.settingsElements.blocks.SettingsBlock
-import view.appInterface.settingsElements.blocks.TitleBarBlock
-import view.appInterface.settingsElements.blocks.WorkAreaBlock
+import view.appInterface.dialogElements.blocks.ColorsBlock
+import view.appInterface.dialogElements.blocks.CommandFieldBlock
+import view.appInterface.dialogElements.blocks.GeneralBlock
+import view.appInterface.dialogElements.blocks.MainScreenBlock
+import view.appInterface.dialogElements.blocks.PerformanceBlock
+import view.appInterface.dialogElements.blocks.SettingsBlock
+import view.appInterface.dialogElements.blocks.TitleBarBlock
+import view.appInterface.dialogElements.blocks.WorkAreaBlock
 
 @Suppress("ktlint:standard:function-naming")
 @Composable
