+++ app/src/main/kotlin/view/appInterface/dialogElements/blocks/ColorsBlock.kt
@@ -1,4 +1,4 @@
-package view.appInterface.settingsElements.blocks
+package view.appInterface.dialogElements.blocks
 
 import androidx.compose.desktop.ui.tooling.preview.Preview
 import androidx.compose.foundation.layout.Column
@@ -37,7 +37,7 @@ import org.coremapx.app.userDirectory.config.ConfigKeys.SURFACE
 import org.coremapx.app.userDirectory.config.ConfigKeys.VERTEX_MAIN_COLOR
 import org.coremapx.app.userDirectory.config.ConfigKeys.WARNING_COLOR
 import view.appInterface.preview.PreviewSurface
-import view.appInterface.settingsElements.lines.ColorPickLine
+import view.appInterface.dialogElements.lines.ColorPickLine
 
 @Suppress("ktlint:standard:function-naming")
 @Composable
