+++ app/src/main/kotlin/view/appInterface/dialogElements/blocks/PerformanceBlock.kt
@@ -1,4 +1,4 @@
-package view.appInterface.settingsElements.blocks
+package view.appInterface.dialogElements.blocks
 
 import androidx.compose.desktop.ui.tooling.preview.Preview
 import androidx.compose.foundation.layout.Column
@@ -14,38 +14,38 @@ import org.coremapx.app.config
 import org.coremapx.app.localization.LocalizationManager
 import org.coremapx.app.localization.objects.LocalizationFormatter
 import org.coremapx.app.theme.AppTheme
-import org.coremapx.app.userDirectory.config.ConfigKeys.TITLE_BAR_HEIGHT
-import org.coremapx.app.userDirectory.config.ConfigKeys.TITLE_BAR_ICON_SIZE
+import org.coremapx.app.userDirectory.config.ConfigKeys.ANIMATION_DURATION
+import org.coremapx.app.userDirectory.config.ConfigKeys.COMMAND_FIELD_SCROLL_DELAY
 import view.appInterface.preview.PreviewSurface
-import view.appInterface.settingsElements.lines.NumberTextFieldLine
+import view.appInterface.dialogElements.lines.NumberTextFieldLine
 
 @Suppress("ktlint:standard:function-naming")
 @Composable
-fun TitleBarBlock(isExpandedSettings: Boolean = config.states.isExpandedSettings.value) {
-    val titleBarHeight by remember { config.states.titleBarHeight }
-    val titleBarIconSize by remember { config.states.titleBarIconSize }
+fun PerformanceBlock(isExpandedSettings: Boolean = config.states.isExpandedSettings.value) {
+    val animationDuration by remember { config.states.animationDuration }
+    val commandFieldScrollDelay by remember { config.states.commandFieldScrollDelay }
 
     Column {
         NumberTextFieldLine(
-            title = LocalizationManager.states.dialogs.titleBarHeight.value,
+            title = LocalizationManager.states.dialogs.performanceAnimationDuration.value,
             valueType = Int::class,
-            value = TextFieldValue("$titleBarHeight"),
-            onValueChange = { config.setValue(TITLE_BAR_HEIGHT, it.text) },
+            value = TextFieldValue("$animationDuration"),
+            onValueChange = { config.setValue(ANIMATION_DURATION, it.text) },
             description =
                 LocalizationFormatter.getStringWithLineBreak(
-                    startString = LocalizationManager.states.descriptions.descriptionTitleBarHeight.value,
+                    startString = LocalizationManager.states.descriptions.descriptionAnimationDuration.value,
                 ),
             isExpanded = isExpandedSettings,
         )
         Spacer(Modifier.height(8.dp))
         NumberTextFieldLine(
-            title = LocalizationManager.states.dialogs.titleBarIconSize.value,
+            title = LocalizationManager.states.dialogs.performanceCommandFieldScrollDelay.value,
             valueType = Int::class,
-            value = TextFieldValue("$titleBarIconSize"),
-            onValueChange = { config.setValue(TITLE_BAR_ICON_SIZE, it.text) },
+            value = TextFieldValue("$commandFieldScrollDelay"),
+            onValueChange = { config.setValue(COMMAND_FIELD_SCROLL_DELAY, it.text) },
             description =
                 LocalizationFormatter.getStringWithLineBreak(
-                    startString = LocalizationManager.states.descriptions.descriptionTitleBarIconSize.value,
+                    startString = LocalizationManager.states.descriptions.descriptionCommandFieldScrollDelay.value,
                 ),
             isExpanded = isExpandedSettings,
         )
@@ -55,8 +55,8 @@ fun TitleBarBlock(isExpandedSettings: Boolean = config.states.isExpandedSettings
 @Suppress("ktlint:standard:function-naming")
 @Preview
 @Composable
-private fun PreviewTitleBarBlock() {
+private fun PreviewPerformanceBlock() {
     AppTheme {
-        PreviewSurface(content = { TitleBarBlock() })
+        PreviewSurface(content = { PerformanceBlock() })
     }
 }
