+++ app/src/main/kotlin/view/appInterface/dialogElements/description/SettingsDescriptionIconButton.kt
@@ -1,4 +1,4 @@
-package view.appInterface.settingsElements.description
+package view.appInterface.dialogElements.description
 
 import androidx.compose.desktop.ui.tooling.preview.Preview
 import androidx.compose.foundation.layout.Column
