+++ app/src/main/kotlin/view/appInterface/dialogElements/lines/ColorPickLine.kt
@@ -1,4 +1,4 @@
-package view.appInterface.settingsElements.lines
+package view.appInterface.dialogElements.lines
 
 import androidx.compose.desktop.ui.tooling.preview.Preview
 import androidx.compose.foundation.layout.Arrangement
@@ -26,8 +26,8 @@ import extensions.border
 import org.coremapx.app.theme.AppTheme
 import view.appInterface.button.ColorPickerButton
 import view.appInterface.preview.PreviewSurface
-import view.appInterface.settingsElements.description.SettingsDescriptionIconButton
-import view.appInterface.settingsElements.description.SettingsDescriptionText
+import view.appInterface.dialogElements.description.SettingsDescriptionIconButton
+import view.appInterface.dialogElements.description.SettingsDescriptionText
 
 @Suppress("ktlint:standard:function-naming")
 @Composable
