+++ app/src/main/kotlin/view/appInterface/dialogElements/lines/NumberTextFieldLine.kt
@@ -1,4 +1,4 @@
-package view.appInterface.settingsElements.lines
+package view.appInterface.dialogElements.lines
 
 import androidx.compose.desktop.ui.tooling.preview.Preview
 import androidx.compose.foundation.interaction.MutableInteractionSource
@@ -40,8 +40,8 @@ import model.result.Result
 import org.coremapx.app.localization.LocalizationManager
 import org.coremapx.app.theme.AppTheme
 import view.appInterface.preview.PreviewSurface
-import view.appInterface.settingsElements.description.SettingsDescriptionIconButton
-import view.appInterface.settingsElements.description.SettingsDescriptionText
+import view.appInterface.dialogElements.description.SettingsDescriptionIconButton
+import view.appInterface.dialogElements.description.SettingsDescriptionText
 import view.appInterface.textField.CustomTextField
 import kotlin.reflect.KClass
